+++ CHANGELOG.md
@@ -26,6 +26,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Default paraphase igv.js track settings: color by YC, squish, extend visibility window, auto-height and place last in view. (#5724)
 - Chanjo2 requests are sent with OIDC id token, if available (#5532)
 - Refactored and simplified SV, cancer SV and MEI variants views (#5701)
+- Replaced deprecated `flask-ldapconn` library with `flask-ldap3-login`. Modified LDAP login module accordingly (#5719)
 ### Fixed
 - Typo in PR template (#5682)
 - Highlight affected individuals/samples on `GT call` tables (#5682)